backend/requirements.txt
django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
channels==4.3.1
